feat:Adding template for the app, with a basic camera
This will be a basic app that allows users to click a photo or record a video and then save it to the cloud using --not decided yet-- service. It will also have a basic gallery view, to allow users to see, which of their photos or videos have been uploaded. Right now it just has feature to open a camera view.