feat: Added Kinvey Client and Tablayout for displaying images and videos
Downloaded and add Kinvey Client and instantiated it by creating a kinvey.properties file in asset folder. This will later be used for uploading images and videos to the cloud. Also made the home screen a tab layout adding seprate fragments for displaying images and videos, resulting in much cleaner and organised code as well as better experience for the user.

Next up building the image gallery.